# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraApi",
        "description": "LaraApi project API documents",
        "contact": {
            "email": "russell57260620@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/resumetag": {
            "get": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "index",
                "description": "Display a listing of the resume_tag table.",
                "operationId": "getResumeTagIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "store",
                "description": "Store a newly created resume-tag map data in resume_tag table.",
                "operationId": "postResumeTagStore",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "index",
                "description": "Display a listing of the resume_tag table.",
                "operationId": "headResumeTagIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
        "/api/resumetag/{id}": {
            "get": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "show",
                "description": "Display the specified resume-tag map data with resumetag id.",
                "operationId": "getResumeTagShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ResumeTag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "update",
                "description": "Update the specified resume-tag map data in resume_tag table.",
                "operationId": "putResumeTagUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ResumeTag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "destroy",
                "description": "Remove the specified resume-tag map data from resume_tag table.",
                "operationId": "deleteResumeTagDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ResumeTag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "show",
                "description": "Display the specified resume-tag map data with resumetag id.",
                "operationId": "headResumeTagShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ResumeTag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "ResumeTag"
                ],
                "summary": "update",
                "description": "Update the specified resume-tag map data in resume_tag table.",
                "operationId": "patchResumeTagUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ResumeTag ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}